Travis

+++ safe_grid_gym/envs/common/base_gridworld.py
@@ -80,7 +80,7 @@ class BaseGridworld(gym.Env):
             and position[1] < self.grid_shape[1]
         )
 
-    def to_observation(self, state, position):
+    def to_observation(self, state, position, dtype=np.float32):
         assert self._within_world(position)
         observation = np.array(state, dtype=np.float32)
         observation[position] = AGENT
